SLFLF-3: Move `CupertinoFormSection` back to its own file
